+++ assets/src/hooks/SortableAssocs/index.js
@@ -9,7 +9,9 @@ export default app => ({
     this.sortableParams = this.el.dataset.sortableParams
     this.sortableBinaryKeys = this.el.dataset.sortableBinaryKeys
     this.sortableDispatchEvent = this.el.dataset.sortableDispatchEvent
+    this.sortableDispatchEventTargetId = this.el.dataset.sortableDispatchEventTargetId
     this.sortablePushEvent = this.el.dataset.sortablePushEvent
+    this.sortableFilter = this.el.dataset.sortableFilter
 
     let sorter = new Sortable(this.el, {
       group: this.sortableId,
@@ -18,14 +20,19 @@ export default app => ({
       draggable: this.sortableSelector || '.draggable',
       ghostClass: 'is-sorting',
       handle: this.handle,
+      filter: this.sortableFilter,
       swapThreshold: 0.5,
       forceFallback: true,
       onEnd: e => {
         if (this.sortableDispatchEvent) {
-          this.el
-            .closest('form')
-            .querySelector('input')
-            .dispatchEvent(new Event('input', { bubbles: true }))
+          let target
+          if (this.sortableDispatchEventTargetId) {
+            // TODO: set debounce to 0, then back again after dispatching event
+            target = document.getElementById(this.sortableDispatchEventTargetId)
+          } else {
+            target = this.el.closest('form').querySelector('input')
+          }
+          target.dispatchEvent(new Event('input', { bubbles: true }))
         }
 
         if (this.sortablePushEvent) {